--- src/routes/linkedin.svelte
@@ -1,8 +1,0 @@
-<script context="module">
-	export async function load() {
-		return {
-			status: 302,
-			redirect: 'https://linkedin.com/in/alex-y-wang'
-		};
-	}
-</script>
