+++ src/routes/links/+page.server.js
@@ -1,0 +1,24 @@
+export async function load() {
+    // basically a dummy function that stores all our links
+    // can put a "name", "url", and a (font awesome) "icon"
+    return {
+    links: [
+        {
+            name: "Personal Site",
+            url: "/"
+        },
+        {
+            name: "Portfolio",
+            url: "/portfolio"
+        },
+        {
+            name: "Unixfy.net",
+            url: "https://unixfy.net"
+        },
+        {
+            name: "Blog",
+            url: "https://blog.unixfy.net"
+        }
+    ]
+}
+}
