+++ src/routes/links/+page.svelte
@@ -1,6 +1,8 @@
 <script>
+    throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
     import { browser } from '$app/env';
-    import { title } from "../stores";
+    import { title } from "../../stores";
     import {page} from "$app/stores";
 	import SocialGrid from '$components/SocialGrid.svelte';
