--- src/routes/portfolio.svelte
@@ -1,8 +1,0 @@
-<script context="module">
-	export async function load() {
-		return {
-			status: 302,
-			redirect: '/projects'
-		};
-	}
-</script>
