+++ src/routes/project/hideout-codes-bot/+page.md
